refactor: share the boot memory log between both entry points
logRuntimeMemory was copy-pasted into index.ts and scheduler.ts. Move it
into logging/runtimeMemory.ts as one exported function taking the logger
and the service name, and drop the inline comments that restated the
pull request body.
